+++ pcs/lib/resource_agent/types.py
@@ -57,6 +57,30 @@ class ResourceAgentActionOcf1_0:  # pylint: disable=invalid-name
 
 
 @dataclass(frozen=True)
+class ResourceAgentActionOcf1_1:  # pylint: disable=invalid-name
+    # pylint: disable=too-many-instance-attributes
+
+    # (start, stop, promote...), mandatory by both OCF 1.0 and 1.1
+    name: str
+    # mandatory by both OCF 1.0 and 1.1, sometimes not defined by agents
+    timeout: Optional[str]
+    # optional by both OCF 1.0 and 1.1
+    interval: Optional[str]
+    # optional by OCF 1.1
+    role: Optional[str]
+    # OCF name: 'start-delay', optional by both OCF 1.0 and 1.1
+    start_delay: Optional[str]
+    # optional by both OCF 1.0 and 1.1
+    depth: Optional[str]
+    # not allowed by any OCF, defined in OCF 1.0 agents anyway, most probably
+    # will be used in OCF 1.1 agents as well as it holds important information
+    automatic: Optional[str]
+    # not allowed by any OCF, defined in OCF 1.0 agents anyway, most probably
+    # will be used in OCF 1.1 agents as well as it holds important information
+    on_target: Optional[str]
+
+
+@dataclass(frozen=True)
 class ResourceAgentParameterOcf1_0:  # pylint: disable=invalid-name
     # pylint: disable=too-many-instance-attributes
 
@@ -83,12 +107,52 @@ class ResourceAgentParameterOcf1_0:  # pylint: disable=invalid-name
 
 
 @dataclass(frozen=True)
+class ResourceAgentParameterOcf1_1:  # pylint: disable=invalid-name
+    # pylint: disable=too-many-instance-attributes
+
+    # name of the parameter
+    name: str
+    # short description
+    shortdesc: Optional[str]
+    # long description
+    longdesc: Optional[str]
+    # data type of the parameter
+    type: str
+    # default value of the parameter
+    default: Optional[str]
+    # allowed values, only defined if type == 'select'
+    enum_values: Optional[List[str]]
+    # is this a required parameter?
+    required: Optional[str]
+    # is this parameter deprecated?
+    deprecated: bool
+    # list of parameters deprecating this one
+    deprecated_by: List[str]
+    # text describing / explaining the deprecation
+    deprecated_desc: Optional[str]
+    # should the parameter's value be unique across same agent resources?
+    # OCF 1.1 defines "unique" as well, but it is deprecated and we ignore it
+    unique_group: Optional[str]
+    # changing this parameter's value triggers a reload instead of a restart
+    reloadable: Optional[str]
+
+
+@dataclass(frozen=True)
 class ResourceAgentMetadataOcf1_0:  # pylint: disable=invalid-name
     name: ResourceAgentName
     shortdesc: Optional[str]
     longdesc: Optional[str]
     parameters: List[ResourceAgentParameterOcf1_0]
     actions: List[ResourceAgentActionOcf1_0]
+
+
+@dataclass(frozen=True)
+class ResourceAgentMetadataOcf1_1:  # pylint: disable=invalid-name
+    name: ResourceAgentName
+    shortdesc: Optional[str]
+    longdesc: Optional[str]
+    parameters: List[ResourceAgentParameterOcf1_1]
+    actions: List[ResourceAgentActionOcf1_1]
 
 
 @dataclass(frozen=True)