+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Option `--autodelete` of command `pcs resource move` is fully supported
   ([rhbz#1990784])
+- Support for OCF 1.1 resource and stonith agents ([rhbz#2018969])
 
 ### Fixed
 - Do not show warning that no stonith device was detected and stonith-enabled
@@ -13,8 +14,9 @@
   ([rhbz#1968088])
 
 [ghpull#370]: https://github.com/ClusterLabs/pcs/pull/370
-[rhbz#1990784]: https://bugzilla.redhat.com/show_bug.cgi?id=1990784
 [rhbz#1968088]: https://bugzilla.redhat.com/show_bug.cgi?id=1968088
+[rhbz#1990784]: https://bugzilla.redhat.com/show_bug.cgi?id=1990784
+[rhbz#2018969]: https://bugzilla.redhat.com/show_bug.cgi?id=2018969
 
 
 ## [0.10.11] - 2021-10-05